+++ gtkwave3/src/helpers/vcd2vzt.c
@@ -49,7 +49,7 @@ int deadcnt=0;
 int ziptype=0;
 
 int opt_depth = 4;
-unsigned long long opt_break_size = 0;
+uint64_t opt_break_size = 0;
 int opt_maxgranule = 8;
 int opt_twostate = 0;
 int opt_rle = 0;
@@ -1963,7 +1963,7 @@ while (1)
 			break;
 
 		case 'b':
-			sscanf(optarg, "%llu", &opt_break_size);
+			sscanf(optarg, "%"SCNu64, &opt_break_size);
 			errno = 0;
 			break;
